+++ admin/tabs/activity/clienterrors/ClientErrorsPanel.js
@@ -7,8 +7,8 @@
 import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
 import {creates, hoistCmp} from '@xh/hoist/core';
 import {button, buttonGroup, colChooserButton, exportButton} from '@xh/hoist/desktop/cmp/button';
-import {dateInput} from '@xh/hoist/desktop/cmp/input';
 import {filterChooser} from '@xh/hoist/desktop/cmp/filter';
+import {dateInput} from '@xh/hoist/desktop/cmp/input';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {toolbar, toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
 import {Icon} from '@xh/hoist/icon';
@@ -16,7 +16,8 @@ import {LocalDate} from '@xh/hoist/utils/datetime';
 import {clientErrorDetail} from './ClientErrorDetail';
 import {ClientErrorsModel} from './ClientErrorsModel';
 
-export const clientErrorsView = hoistCmp.factory({
+
+export const clientErrorsPanel = hoistCmp.factory({
     model: creates(ClientErrorsModel),
 
     render() {