+++ admin/tabs/monitor/MemoryMonitorModel.js
@@ -1,16 +1,23 @@
+/*
+ * This file belongs to Hoist, an application development toolkit
+ * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
+ *
+ * Copyright © 2021 Extremely Heavy Industries Inc.
+ */
 import {ChartModel} from '@xh/hoist/cmp/chart';
-import {dateTimeCol, GridModel} from '@xh/hoist/cmp/grid';
-import {HoistModel, XH} from '@xh/hoist/core';
-import {fmtTime, numberRenderer} from '@xh/hoist/format';
+import {GridModel} from '@xh/hoist/cmp/grid';
+import {HoistModel, XH, managed} from '@xh/hoist/core';
+import {fmtTime} from '@xh/hoist/format';
 import {checkMinVersion} from '@xh/hoist/utils/js';
 import {forOwn, sortBy} from 'lodash';
+import * as MCol from './MonitorColumns';
 
 export class MemoryMonitorModel extends HoistModel {
 
     /** @member {GridModel} */
-    gridModel;
+    @managed gridModel;
     /** @member {ChartModel} */
-    chartModel;
+    @managed chartModel;
 
     get minVersionWarning() {
         const minVersion = '8.7.0',
@@ -20,30 +27,20 @@ export class MemoryMonitorModel extends HoistModel {
 
     constructor() {
         super();
-        const mbCol = {width: 150, renderer: numberRenderer({precision: 2, useCommas: true})},
-            pctCol = {width: 150, renderer: numberRenderer({precision: 2, useCommas: true, label: '%'})};
 
         this.gridModel = new GridModel({
             sortBy: 'timestamp|desc',
             enableExport: true,
-            store: {
-                idSpec: 'timestamp',
-                fields: [
-                    {name: 'timestamp', type: 'date'},
-                    {name: 'totalHeapMb', type: 'number', displayName: 'Total (mb)'},
-                    {name: 'maxHeapMb', type: 'number', displayName: 'Max (mb)'},
-                    {name: 'freeHeapMb', type: 'number', displayName: 'Free (mb)'},
-                    {name: 'usedHeapMb', type: 'number', displayName: 'Used (mb)'},
-                    {name: 'usedPctTotal', type: 'number', displayName: 'Used (pct Total)'}
-                ]
-            },
+            filterModel: true,
+            store: {idSpec: 'timestamp'},
+            colDefaults: {filterable: true},
             columns: [
-                {field: 'timestamp', ...dateTimeCol},
-                {field: 'totalHeapMb', ...mbCol},
-                {field: 'maxHeapMb', ...mbCol},
-                {field: 'freeHeapMb', ...mbCol},
-                {field: 'usedHeapMb', ...mbCol},
-                {field: 'usedPctTotal', ...pctCol}
+                MCol.timestamp,
+                MCol.totalHeapMb,
+                MCol.maxHeapMb,
+                MCol.freeHeapMb,
+                MCol.usedHeapMb,
+                MCol.usedPctTotal
             ]
         });
 
@@ -155,5 +152,4 @@ export class MemoryMonitorModel extends HoistModel {
             XH.handleException(e);
         }
     }
-
 }