+++ cmp/layout/TileFrame.js
@@ -7,7 +7,7 @@
 import {hoistCmp, useLocalModel, HoistModel} from '@xh/hoist/core';
 import {frame, box} from '@xh/hoist/cmp/layout';
 import {useOnResize} from '@xh/hoist/utils/react';
-import {useState} from 'react';
+import {useState, useLayoutEffect} from 'react';
 import {minBy, isEqual} from 'lodash';
 import composeRefs from '@seznam/compose-react-refs';
 import PT from 'prop-types';
@@ -37,6 +37,7 @@ export const [TileFrame, tileFrame] = hoistCmp.withFactory({
         maxTileWidth,
         minTileHeight,
         maxTileHeight,
+        onLayoutChange,
         ...props
     }, ref) {
         const localModel = useLocalModel(() => new LocalModel()),
@@ -56,6 +57,11 @@ export const [TileFrame, tileFrame] = hoistCmp.withFactory({
             minTileHeight,
             maxTileHeight
         });
+
+        useLayoutEffect(
+            () => onLayoutChange?.(localModel.layout),
+            [onLayoutChange, localModel.layout]
+        );
 
         ref = composeRefs(
             ref,
@@ -98,7 +104,13 @@ TileFrame.propTypes = {
     minTileHeight: PT.number,
 
     /** Max tile height (in px).*/
-    maxTileHeight: PT.number
+    maxTileHeight: PT.number,
+
+    /**
+     * Callback triggered when the layout configuration changes.
+     * Receives the layout object {rows, cols, tileWidth, tileHeight} as its sole argument.
+     */
+    onLayoutChange: PT.func
 };
 
 class LocalModel extends HoistModel {
@@ -109,7 +121,10 @@ class LocalModel extends HoistModel {
     setParams(params) {
         if (isEqual(params, this.params)) return;
         this.params = params;
-        this.layout = this.createLayout();
+
+        const layout = this.createLayout();
+        if (isEqual(layout, this.layout)) return;
+        this.layout = layout;
     }
 
     getTileStyle(idx) {